+++ lib/src/utils/constants.dart
@@ -21,6 +21,7 @@ const redditBaseURL = 'https://reddit.com/r/';
 const telegramBaseURL = 'https://t.me/';
 const facebookMessengerURL = 'https://m.me/';
 const whatsappURL = 'https://wa.me/';
+const googleDriveURL = 'https://docs.google.com/viewer?url=';
 
 const SELECTED_LANGUAGE_CODE = 'selected_language_code';
 const THEME_MODE_INDEX = 'theme_mode_index';