+++ lib/src/utils/date_time_utils.dart
@@ -5,13 +5,28 @@ extension DateTimeExt on DateTime {
   String get timeAgo => formatTime(this.millisecondsSinceEpoch);
 
   /// Returns true if given date is today
-  bool get isToday => DateTime.now().day == this.day;
+  bool get isToday {
+    final now = DateTime.now();
+    final today = DateTime(now.year, now.month, now.day);
+
+    return DateTime(this.year, this.month, this.day) == today;
+  }
 
   /// Returns true if given date is yesterday
-  bool get isYesterday => DateTime.now().subtract(1.days).day == this.day;
+  bool get isYesterday {
+    final now = DateTime.now();
+    final yesterday = DateTime(now.year, now.month, now.day - 1);
+
+    return DateTime(now.year, now.month, now.day - 1) == yesterday;
+  }
 
   /// Returns true if given date is tomorrow
-  bool get isTomorrow => DateTime.now().add(1.days).day == this.day;
+  bool get isTomorrow {
+    final now = DateTime.now();
+    final yesterday = DateTime(now.year, now.month, now.day - 1);
+
+    return DateTime(now.year, now.month, now.day + 1) == yesterday;
+  }
 }
 
 /// return current time in milliseconds