+++ lib/src/widgets/AppButton.dart
@@ -1,7 +1,5 @@
 import 'package:flutter/material.dart';
-import 'package:flutter/rendering.dart';
 import 'package:nb_utils/nb_utils.dart';
-import 'package:nb_utils/src/utils/text_styles.dart';
 
 /// Default App Button
 class AppButton extends StatefulWidget {