+++ lib/src/widgets/AppTextField.dart
@@ -1,7 +1,6 @@
 import 'package:flutter/material.dart';
 import 'package:flutter/services.dart';
 import 'package:nb_utils/nb_utils.dart';
-import 'package:nb_utils/src/utils/text_styles.dart';
 
 enum TextFieldType {
   EMAIL,